requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.16
openai>=1.3.5
tiktoken>=0.5.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
numpy>=1.24.3
pandas>=2.0.3
pydantic>=2.0.3
pypdf>=3.15.1
chromadb>=0.4.18
torch>=2.0.1
transformers>=4.35.2
accelerate>=0.20.3
bitsandbytes>=0.41.1
flash-attn>=2.3.3 